+++ src/lighting.js
@@ -47,6 +47,8 @@ const { dmx_scene_command,
           = require('./lightingDmx');
 const { shelly_command }                        // Shelly device module
            = require('./lightingShelly.js');
+const { wled_command }                          // WLED device module
+           = require('./lightingWled.js');
 
 
 // Load variables
@@ -91,6 +93,17 @@ function lighting_device_command (operation_num, device, lighting_command) {
         case 'shelly':
           console.log("%d - Turning Shelly %s to %s",operation_num, lights_new[i].address, lighting_command);
           shelly_command(operation_num, lights_new[i].address, lighting_command);
+          if(lighting_command=="OFF"){
+              lights_new[i].checked = false;
+              lights_new[i].status = 0;
+          } else if(lighting_command=="ON"){
+              lights_new[i].checked = true;
+              lights_new[i].status = 100;
+          }
+          break;
+        case 'wled':
+          console.log("%d - Turning WLED %s to %s",operation_num, lights_new[i].address, lighting_command);
+          wled_command(operation_num, lights_new[i].address, lighting_command);
           if(lighting_command=="OFF"){
               lights_new[i].checked = false;
               lights_new[i].status = 0;